[snapshot]: remove wrong assumption of read buffer size
Similar to ag/34109605, this cl fixes the wrong
assumption that len is less than max value of signed
32bit integer.
That assumption lead to a real crash, when
read len=2,258,424,990 > 2,147,483,647(a.k.a., 2**31-1)

Bug: 417788798

+++ android-qemu2-glue/base/files/QemuFileStream.cpp
@@ -43,9 +43,7 @@ void* QemuFileStream::getProtobuf() {
 ssize_t QemuFileStream::read(void* buffer, size_t len) {
     DCHECK(static_cast<ssize_t>(len) >= 0);
     return static_cast<ssize_t>(
-            qemu_get_buffer(mFile,
-                            static_cast<uint8_t*>(buffer),
-                            static_cast<int>(len)));
+            qemu_get_buffer(mFile, static_cast<uint8_t*>(buffer), len));
 }
 
 ssize_t QemuFileStream::write(const void* buffer, size_t len) {
